exam creation updated
