+++ src/main/java/com/testify/Testify_Backend/model/Exam.java
@@ -5,11 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.SuperBuilder;
 
 import java.util.Set;
 @Entity
 @Getter
 @Setter
+@SuperBuilder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Exam {
@@ -19,23 +21,22 @@ public class Exam {
 
     @Column(nullable = false)
     private String title;
+
+    @ManyToOne
+    private ExamSetter examSetter;
+
+    @ManyToOne
+    private Organization organization;
+
     private String description;
     private String instructions;
     private int duration;
     private int totalMarks;
     private int passMarks;
 
-    @ManyToOne
-    private ExamSetter examSetter;
-    @ManyToOne
-    private Organization organization;
+
 
     @OneToMany(mappedBy = "exam", cascade = CascadeType.ALL)
     private Set<Question> questions;
-
-
-
-//    @OneToMany(mappedBy = "enrolled_exam", cascade = CascadeType.ALL)
-//    private Set<User> users;
 
 }
