+++ src/main/java/com/testify/Testify_Backend/repository/AttendeeRepository.java
@@ -1,10 +1,10 @@
 package com.testify.Testify_Backend.repository;
 
-import com.testify.Testify_Backend.model.Attendee;
+import com.testify.Testify_Backend.model.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AttendeeRepository extends JpaRepository<Attendee, Long> {
-    Optional<Attendee> findByEmail(String email);
+public interface AttendeeRepository extends JpaRepository<Candidate, Long> {
+    Optional<Candidate> findByEmail(String email);
 }
